[user][fix] Controller 테스트 코드 작성 (Mock 사용)
- Controller 테스트 코드 작성
- User -> Account Refactor
