--- src/test/java/com/bootproj/pmcweb/web/UserControllerTest.java
@@ -1,12 +1,0 @@
-package com.bootproj.pmcweb.web;
-
-
-import com.bootproj.pmcweb.Controller.AccountController;
-import org.junit.runner.RunWith;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.context.junit4.SpringRunner;
-
-@RunWith(SpringRunner.class)
-@WebMvcTest(AccountController.class)
-public class UserControllerTest {
-}
